+++ tests/continual/test_container.py
@@ -121,17 +121,17 @@ def test_sequential_with_TensorPlaceholder():
     assert torch.allclose(target[:, :, :-1], out_cleaned)
 
 
-def test_sum_aggregation():
+def test_sum_reduce():
     ones = torch.ones((1, 2, 4, 3, 3))
     twos = torch.ones((1, 2, 4, 3, 3)) * 2
-    res = co.container.parallel_sum([ones, ones])
+    res = co.container.reduce_sum([ones, ones])
     assert torch.allclose(res, twos)
 
 
-def test_concat_aggregation():
+def test_concat_reduce():
     ones = torch.ones((1, 2, 4, 3, 3))
     twos = torch.ones((1, 2, 4, 3, 3)) * 2
-    res = co.container.parallel_concat([ones, twos])
+    res = co.container.reduce_concat([ones, twos])
     assert res.shape == (1, 4, 4, 3, 3)
     assert torch.allclose(res[:, :2], ones)
     assert torch.allclose(res[:, 2:], twos)
@@ -167,7 +167,7 @@ def test_residual():
     assert torch.allclose(out_last, target[:, :, -2])
 
 
-def test_parallel():
+def test_broadcast_reduce():
     input = torch.arange(7, dtype=torch.float).reshape((1, 1, 7))
 
     c5 = co.Conv1d(1, 1, kernel_size=5, padding=2, bias=False)
@@ -176,12 +176,12 @@ def test_parallel():
     torch.nn.init.ones_(c5.weight)
     torch.nn.init.ones_(c3.weight)
     torch.nn.init.ones_(c1.weight)
-    par = co.Parallel(OrderedDict([("c5", c5), ("c3", c3), ("c1", c1)]))
+    par = co.BroadcastReduce(OrderedDict([("c5", c5), ("c3", c3), ("c1", c1)]))
 
     assert par.stride == 1
     assert par.delay == 2
     assert par.padding == 2
-    assert "Parallel(" in par.__repr__() and "aggregation_fn=" in par.__repr__()
+    assert "BroadcastReduce(" in par.__repr__() and "reduce=" in par.__repr__()
 
     # forward
     out_all = par.forward(input)
@@ -231,7 +231,7 @@ def test_flat_state_dict():
     assert set(sd_no_flat) == {"c1.weight", "c1.bias"}
 
     # A nested example:
-    nested = co.Parallel(seq_to_flatten, seq_not_to_flatten)
+    nested = co.BroadcastReduce(seq_to_flatten, seq_not_to_flatten)
     sd = nested.state_dict()
     assert set(sd) == {"0.0.weight", "0.0.bias", "1.c1.weight", "1.c1.bias"}
 
@@ -239,7 +239,7 @@ def test_flat_state_dict():
     assert set(sd_flat) == {"weight", "bias", "c1.weight", "c1.bias"}
 
     # >> Part 2: Load flat state dict
-    nested_new = co.Parallel(
+    nested_new = co.BroadcastReduce(
         co.Sequential(nn.Conv1d(1, 1, 3)),
         co.Sequential(OrderedDict([("c1", nn.Conv1d(1, 1, 3))])),
     )
@@ -324,3 +324,88 @@ def test_conditional_delay():
     assert mod.delay == 3
     assert mod._modules["0"].delay == 3
     assert mod._modules["1"].delay == 3
+
+
+def test_broadcast():
+    x = 42
+
+    mod = co.Broadcast(2)
+    assert mod.delay == 0
+    assert mod.forward(x) == [x, x]
+    assert mod.forward_step(x) == [x, x]
+    assert mod.forward_steps(x) == [x, x]
+
+
+def test_parallel():
+    x = torch.randn((1, 1, 3))
+    xx = [x, x]
+
+    c3 = co.Conv1d(1, 1, kernel_size=3, padding=1, bias=False)
+    c1 = co.Conv1d(1, 1, kernel_size=1, padding=0, bias=False)
+
+    par = co.Parallel(OrderedDict([("c3", c3), ("c1", c1)]))
+    assert par.delay == 1
+    assert par.padding == 1
+    assert par.stride == 1
+
+    o1 = par.forward(xx)
+    assert torch.equal(c3.forward(x), o1[0])
+    assert torch.equal(c1.forward(x), o1[1])
+
+    o2 = par.forward_steps(xx, pad_end=True, update_state=False)
+    assert torch.equal(c3.forward_steps(x, pad_end=True), o2[0])
+    assert torch.equal(c1.forward_steps(x, pad_end=True), o2[1])
+
+    par.clean_state()
+    par.forward_step([x[:, :, 0], x[:, :, 0]], update_state=True)
+    o3 = par.forward_step([x[:, :, 1], x[:, :, 1]], update_state=False)
+    assert torch.equal(c3.forward_step(x[:, :, 1]), o3[0])
+    assert torch.equal(c1.forward_step(x[:, :, 0]), o3[1])  # x[:,:,0] due to auto delay
+
+
+def test_reduce():
+    x = torch.tensor([[[1.0, 2.0]]])
+    xx = [x, x]
+
+    mod = co.Reduce("sum")
+    assert mod.delay == 0
+    assert torch.equal(mod.forward(xx), torch.tensor([[[2.0, 4.0]]]))
+    assert torch.equal(mod.forward_steps(xx), torch.tensor([[[2.0, 4.0]]]))
+    assert torch.equal(
+        mod.forward_step([x[:, :, 0], x[:, :, 0]]), torch.tensor([[2.0]])
+    )
+
+
+def test_parallel_sequential():
+    x = torch.arange(7, dtype=torch.float).reshape((1, 1, 7))
+
+    # Test two equivalent implementations
+    # First
+    c5 = co.Conv1d(1, 1, kernel_size=5, padding=2, bias=False)
+    c3 = co.Conv1d(1, 1, kernel_size=3, padding=1, bias=False)
+    c1 = co.Conv1d(1, 1, kernel_size=1, padding=0, bias=False)
+    torch.nn.init.ones_(c5.weight)
+    torch.nn.init.ones_(c3.weight)
+    torch.nn.init.ones_(c1.weight)
+
+    mod1 = co.BroadcastReduce(c5, c3, c1, reduce="sum")
+
+    # Second
+    c5 = co.Conv1d(1, 1, kernel_size=5, padding=2, bias=False)
+    c3 = co.Conv1d(1, 1, kernel_size=3, padding=1, bias=False)
+    c1 = co.Conv1d(1, 1, kernel_size=1, padding=0, bias=False)
+    torch.nn.init.ones_(c5.weight)
+    torch.nn.init.ones_(c3.weight)
+    torch.nn.init.ones_(c1.weight)
+
+    mod2 = co.Sequential(
+        co.Broadcast(),  # Sequential can infer broadcast dimensions
+        co.Parallel(c5, c3, c1),
+        co.Reduce("sum"),
+    )
+
+    # Compare
+    o1 = mod1.forward(x)
+    o2 = mod2.forward(x)
+
+    assert torch.equal(o1, o2)